requirements.txt
streamlit
mysql-connector
pandas